Started working on Pause Menu
Added a half functioning pause menu to Main_Jacob. When you click the pause button, it brings up a menu. The menu isn't interactable atm which is the problem, but it saves the state the player was in, so when you hit the pause button again it resumes. The resume button will use the same code as pause.